refactor(meal-form): remove deprecated meal record APIs and streamline fetching logic
This commit removes several unused API files related to meal records, including `addMealRecord`, `fetchMealRecords`, and `updMealRecord`. The logic for fetching meal records has been centralized into utility functions (`fetchAndSetMealRecords` and `fetchAndSetWeeklyMealRecords`) to reduce redundancy and improve maintainability.

The refactoring also enhances the consistency of data handling when saving and updating meal records. These changes contribute to a cleaner codebase and set the foundation for future enhancements in the meal tracking feature.
