--- src/features/meal-form/api/index.ts
@@ -1,3 +1,0 @@
-export * from './addMealRecord'
-export * from './fetchMealRecords'
-export * from './updMealRecord'
