+++ src/features/weekly-nutritions/components/WeeklyNutritions.tsx
@@ -3,11 +3,11 @@ import { useEffect } from 'react'
 
 import { Nutritions } from '../../../features/nutritions'
 import { useCurrentDateStore, useWeeklyMealRecordsStore } from '../../../stores'
-import { fetchWeeklyMealRecords } from '../api/fetchWeeklyMealRecords'
 import {
   countWeeklyDateWithFoods,
   createAvgWeekNutritionValues,
   createWeekDateString,
+  fetchAndSetWeeklyMealRecords,
 } from '../utils'
 
 export function WeeklyNutritions() {
@@ -15,20 +15,18 @@ export function WeeklyNutritions() {
   const { weeklyMealRecords, setWeeklyMealRecords } =
     useWeeklyMealRecordsStore()
 
-  const { currentDateString, prevWeekDateString } =
-    createWeekDateString(currentDate)
-
   const weeklyDateWithFoodsCount = countWeeklyDateWithFoods(weeklyMealRecords)
-
   const {
     avgWeeklyCalories,
     avgWeeklyProtein,
     avgWeeklyFat,
     avgWeeklyCarbohydrates,
   } = createAvgWeekNutritionValues(weeklyMealRecords, weeklyDateWithFoodsCount)
+  const { currentDateString, prevWeekDateString } =
+    createWeekDateString(currentDate)
 
   useEffect(() => {
-    fetchWeeklyMealRecords(
+    fetchAndSetWeeklyMealRecords(
       currentDateString,
       prevWeekDateString,
       setWeeklyMealRecords
